Split user-role.dto into two files
