+++ services/121-service/src/user/dto/create-user-role.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsArray, IsNotEmpty, IsOptional, IsString } from 'class-validator';
+import { IsNotEmpty, IsString } from 'class-validator';
 
 import { PermissionEnum } from '@121-service/src/user/enum/permission.enum';
 
@@ -22,24 +22,4 @@ export class CreateUserRoleDto {
     example: Object.values(PermissionEnum),
   })
   public readonly permissions: PermissionEnum[];
-}
-
-export class UpdateUserRoleDto {
-  @ApiProperty({ example: 'Updated user role label' })
-  @IsString()
-  @IsOptional()
-  public readonly label?: string;
-
-  @ApiProperty({ example: 'Updated user role description' })
-  @IsString()
-  @IsOptional()
-  public readonly description?: string;
-
-  @ApiProperty({
-    enum: PermissionEnum,
-    example: Object.values(PermissionEnum),
-  })
-  @IsArray()
-  @IsOptional()
-  public readonly permissions?: PermissionEnum[];
 }
