+++ services/121-service/src/user/user.service.ts
@@ -22,6 +22,7 @@ import {
 import { changePasswordWithoutCurrentPasswordDto } from '@121-service/src/user/dto/change-password-without-current-password.dto';
 import { CookieSettingsDto } from '@121-service/src/user/dto/cookie-settings.dto';
 import { CreateUsersDto } from '@121-service/src/user/dto/create-user.dto';
+import { CreateUserRoleDto } from '@121-service/src/user/dto/create-user-role.dto';
 import { FindUserReponseDto } from '@121-service/src/user/dto/find-user-response.dto';
 import { GetUserReponseDto } from '@121-service/src/user/dto/get-user-response.dto';
 import { LoginResponseDto } from '@121-service/src/user/dto/login-response.dto';
@@ -30,10 +31,7 @@ import {
   UpdateUserDto,
   UpdateUserPasswordDto,
 } from '@121-service/src/user/dto/update-user.dto';
-import {
-  CreateUserRoleDto,
-  UpdateUserRoleDto,
-} from '@121-service/src/user/dto/user-role.dto';
+import { UpdateUserRoleDto } from '@121-service/src/user/dto/update-user-role.dto';
 import {
   AssignmentResponseDTO,
   UserRoleResponseDTO,
